+++ src/cpp/include/openvino/genai/continuous_batching_pipeline.hpp
@@ -61,9 +61,7 @@ public:
     GenerationHandle add_request(uint64_t request_id, const ov::Tensor& input_ids, const ov::genai::GenerationConfig& sampling_params);
     GenerationHandle add_request(uint64_t request_id, const std::string& prompt, const ov::genai::GenerationConfig& sampling_params);
 
-    std::vector<std::string> get_model_configuration();
-
-    void print_model_configuration();
+    std::string get_model_configuration_string();
 
     void step();
