+++ src/cpp/src/debug_utils.hpp
@@ -33,46 +33,3 @@ void print_tensor(std::string name, ov::Tensor tensor) {
         print_array(tensor.data<bool>(), tensor.get_size());
     }
 }
-
-std::string join(const std::vector<std::string>& listOfStrings, const std::string delimiter) {
-    std::stringstream ss;
-    auto it = listOfStrings.cbegin();
-    if (it == listOfStrings.end()) {
-        return "";
-    }
-    for (; it != (listOfStrings.end() - 1); ++it) {
-        ss << *it << delimiter;
-    }
-    if (it != listOfStrings.end()) {
-        ss << *it;
-    }
-    return ss.str();
-}
-
-template <typename PropertyExtractor>
-static void read_properties(PropertyExtractor&& property_extractor, std::vector<std::string>& output_configuration_values) {
-    auto key = std::string("SUPPORTED_PROPERTIES");  // ov::supported_properties;
-    std::vector<ov::PropertyName> supported_config_keys;
-    try {
-        ov::Any value = property_extractor(key);
-        supported_config_keys = value.as<std::vector<ov::PropertyName>>();
-    } catch (...) {
-        std::cout << "Exception thrown from OpenVINO when requesting model property: " << key << std::endl;
-        return;
-    }
-
-    for (auto& key : supported_config_keys) {
-        if (key == "SUPPORTED_PROPERTIES")
-            continue;
-        std::string value;
-        try {
-            ov::Any param_value = property_extractor(key);
-            value = param_value.as<std::string>();
-        } catch (...) {
-            std::cout << "WARNING: Exception thrown from OpenVINO when requesting model property: " << key << std::endl;
-            continue;
-        }
-        output_configuration_values.emplace_back(join({key, value}, ": "));
-    }
-    std::sort(output_configuration_values.begin(), output_configuration_values.end());
-}
